fixes and refactoring for semsim pattern matching (removes semsim instances per variable assignment)
