+++ graphbrain/patterns/matcher.pyx
@@ -7,9 +7,9 @@ from graphbrain.hypergraph import Hypergraph
 from graphbrain.patterns.argroles import _match_by_argroles
 from graphbrain.patterns.atoms import _matches_atomic_pattern
 from graphbrain.patterns.properties import is_fun_pattern, is_pattern, FUNS
-from graphbrain.patterns.semsim.matching import match_semsim
 from graphbrain.patterns.semsim.types import SEMSIM_FUNS
-from graphbrain.patterns.semsim.instances import SemSimInstance, process_semsim_instance_vars
+from graphbrain.patterns.semsim.matching import match_semsim
+from graphbrain.patterns.semsim.instances import SemSimInstance
 from graphbrain.patterns.utils import _defun_pattern_argroles, _atoms_and_tok_pos
 from graphbrain.patterns.variables import _varname, _assign_edge_to_var
 from graphbrain.utils.lemmas import lemma
@@ -28,11 +28,8 @@ class Matcher:
         self.hg: Hypergraph = hg
         self.skip_semsim: bool = skip_semsim
 
-        self.semsim_instances_mapped: Dict[int, SemSimInstance] = {}  # mapping from instance idx to instance
-        self.semsim_instances_sorted: List[List[SemSimInstance]] = []  # store the sorted instances here
-
+        self.semsim_instances: List[SemSimInstance] = []
         self.results: List[Dict] = self.match(edge, pattern, curvars=curvars, tok_pos=tok_pos)
-        process_semsim_instance_vars(self)
 
     def match(self, edge, pattern, curvars=None, tok_pos=None):
         if curvars is None:
