+++ graphbrain/patterns/semsim/processing.py
@@ -1,38 +1,18 @@
-from typing import Dict, List, Union, TYPE_CHECKING
+from typing import List, Union
 
 from graphbrain.hyperedge import Hyperedge
 from graphbrain.hypergraph import Hypergraph
+from graphbrain.patterns.semsim.instances import SemSimInstance
 from graphbrain.patterns.semsim.references import _get_ref_tok_poses
 from graphbrain.semsim import semsim
 
-if TYPE_CHECKING:
-    from graphbrain.patterns.matcher import Matcher
-    from graphbrain.patterns.semsim.instances import SemSimInstance
 
-
-def filter_results_by_semsim_instances(
-        matcher: 'Matcher',
-        pattern: Hyperedge,
-        edge: Hyperedge,
-        ref_edges: List[Hyperedge] = None,
-):
-    # results_post_semsim: List[Dict] = []
-    # for result, semsim_instances in zip(matcher.results, matcher.semsim_instances_sorted):
-    #     if process_semsim_instances(semsim_instances, pattern, edge, matcher.hg, ref_edges=ref_edges):
-    #         results_post_semsim.append(result)
-    # return results_post_semsim
-
-    return [
-        result for result, semsim_instances in zip(matcher.results, matcher.semsim_instances_sorted)
-        if process_semsim_instances(semsim_instances, pattern, edge, matcher.hg, ref_edges=ref_edges)
-    ]
-
-
-def process_semsim_instances(
-        semsim_instances: List['SemSimInstance'],
+def match_semsim_instances(
+        semsim_instances: List[SemSimInstance],
         pattern: Hyperedge,
         edge: Hyperedge,
         hg: Hypergraph,
+        threshold: float = None,
         ref_words: List[str] = None,
         ref_edges: List[Hyperedge] = None,
 ) -> bool:
@@ -44,12 +24,15 @@ def process_semsim_instances(
         ref_tok_poses_per_ref_edge: List[List[Hyperedge]] = _get_ref_tok_poses(pattern, ref_edges, hg)
 
     for instance_idx, instance in enumerate(semsim_instances):
-        ref_tok_poses: List[Hyperedge] = [
-            ref_tok_poses[instance_idx] for ref_tok_poses in ref_tok_poses_per_ref_edge
-        ] if ref_tok_poses_per_ref_edge else None
+        threshold: float = threshold or instance.threshold
+
+        ref_tok_poses: List[Hyperedge] | None = None
+        if ref_tok_poses_per_ref_edge:
+            ref_tok_poses = [ref_tok_poses[instance_idx] for ref_tok_poses in ref_tok_poses_per_ref_edge]
+
         if not semsim(
             semsim_type=instance.type,
-            threshold=instance.threshold,
+            threshold=threshold,
             cand_word=instance.word,
             ref_words=ref_words,
             cand_edge=edge,
