+++ graphbrain/semsim/__init__.py
@@ -1,2 +1,2 @@
-from .interface import semsim, get_matcher, init_matcher
+from .interface import semsim, get_matcher, init_matcher, DEFAULT_CONFIGS
 from .matcher.matcher import SemSimConfig, SemSimMatcher, SemSimType
